requirements.txt
tensorflow-cpu
streamlit
numpy
pillow